# backend/flights-booking-api/src/FlightsBookingAPI/Controllers/FlightApi.cs
/*
 * FlightBookingAPI
 *
 * API aplikacije za kupovinu avionskih karata
 *
 * The version of the OpenAPI document: 1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using FlightsBookingAPI.Attributes;
using FlightsBookingAPI.Models;

namespace FlightsBookingAPI.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class FlightApiController : ControllerBase
    { 
        /// <summary>
        /// DELETE Flight
        /// </summary>
        /// <remarks>Endpoint za brisanje leta.</remarks>
        /// <param name="flightId">Identifikator leta.</param>
        /// <response code="204">No Content</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found</response>
        /// <response code="500">Internal Server Error</response>
        [HttpDelete]
        [Route("/flights/{FlightId}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteFlightsId")]
        public abstract Task<IActionResult> DeleteFlightsId([FromRoute (Name = "FlightId")][Required]Guid flightId);

        /// <summary>
        /// GET Flights
        /// </summary>
        /// <remarks>Endpoint dobavlja podatke o svim letovima.</remarks>
        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/flights")]
        [ValidateModelState]
        [SwaggerOperation("GetFlights")]
        [SwaggerResponse(statusCode: 200, type: typeof(FlightGetAllResponse), description: "OK")]
        public abstract Task<IActionResult> GetFlights();

        /// <summary>
        /// ACTION SearchFlights
        /// </summary>
        /// <remarks>Pretraga letova po zadatim parametrima.</remarks>
        /// <param name="arrivalPlace">Naziv mesta polaska</param>
        /// <param name="departurePlace">Naziv mesta dolaska</param>
        /// <param name="departureDate">Datum polaska</param>
        /// <param name="availableTickets">Broj dostupnih karata na prodaju</param>
        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/flights/actions/search")]
        [ValidateModelState]
        [SwaggerOperation("GetFlightsActionsSearch")]
        [SwaggerResponse(statusCode: 200, type: typeof(FlightGetAllResponse), description: "OK")]
        public abstract Task<IActionResult> GetFlightsActionsSearch([FromQuery (Name = "arrivalPlace")]string arrivalPlace, [FromQuery (Name = "departurePlace")]string departurePlace, [FromQuery (Name = "departureDate")]string departureDate, [FromQuery (Name = "availableTickets")]decimal? availableTickets);

        /// <summary>
        /// GET Flight
        /// </summary>
        /// <remarks>Endpoint za dobavljanje podataka o odabranom letu.</remarks>
        /// <param name="flightId">Identifikator leta.</param>
        /// <response code="200">OK</response>
        /// <response code="404">Not Found</response>
        /// <response code="500">Internal Server Error</response>
        [HttpGet]
        [Route("/flights/{FlightId}")]
        [ValidateModelState]
        [SwaggerOperation("GetFlightsId")]
        [SwaggerResponse(statusCode: 200, type: typeof(FlightGetResponse), description: "OK")]
        public abstract Task<IActionResult> GetFlightsId([FromRoute (Name = "FlightId")][Required]Guid flightId);

        /// <summary>
        /// POST Flight
        /// </summary>
        /// <remarks>Endpoint za kreiranje novog leta.</remarks>
        /// <param name="flightCreateRequest">Saljemo podatke o mestu i vremenu polaska i dolaska, broju i ceni karata.</param>
        /// <response code="200">OK</response>
        [HttpPost]
        [Route("/flights")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PostFlights")]
        public abstract Task<IActionResult> PostFlights([FromBody]FlightCreateRequest flightCreateRequest);

        /// <summary>
        /// ACTION BuyTicket
        /// </summary>
        /// <remarks>Endpoint za kupovinu karata od strane korisnika.</remarks>
        /// <param name="flightId">Identifikator leta.</param>
        /// <param name="flightBuyTicketsRequest"></param>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request: Please specify the amount of tickets.</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found: Flight not found.</response>
        /// <response code="409">Conflict: No available tickets for choosen flight.</response>
        [HttpPost]
        [Route("/flights/{FlightId}/actions/buy-tickets")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PostFlightsIdActionsBuyTicket")]
        [SwaggerResponse(statusCode: 200, type: typeof(FlightBuyTicketsResponse), description: "OK")]
        public abstract Task<IActionResult> PostFlightsIdActionsBuyTicket([FromRoute (Name = "FlightId")][Required]string flightId, [FromBody]FlightBuyTicketsRequest flightBuyTicketsRequest);
    }
}